[mtouch/mmp] Share Touch code.
